kfk_client/pyproject.toml
[project]
name = "client-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.14",
    "aiokafka>=0.12.0",
    "aiosmtplib>=4.0.0",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi-users[sqlalchemy]>=14.0.1",
    "fastapi[all]>=0.115.11",
    "faststream[rabbit]>=0.5.39",
    "google-api-python-client>=2.166.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "greenlet>=3.1.1",
    "gunicorn>=23.0.0",
    "pyjwt[crypto]>=2.10.1",
    "sendgrid>=6.11.0",
    "sqlalchemy>=2.0.38",
]

[dependency-groups]
dev = ["black>=25.1.0", "pyright>=1.1.396"]
test = ["pytest-asyncio>=0.25.3"]
